move Pregel to o.g.lib
